sinchronize with dev
git-svn-id: http://emulsion.na.infn.it/svn/fedra/trunk@1288 96872706-0045-0410-9534-84be8447c86c

+++ src/libEMC/EdbDataStore.cxx
@@ -473,7 +473,7 @@ void EdbDataStore::SavePlateToRaw(int PID,const char* dir,int id){
     run.GetView()->SetNframes(1,0);
     for(int ns=0;ns<ptop->N();++ns){
       seg=ptop->GetSegment(ns);
-      run.GetView()->AddSegment(seg->X(),seg->Y(),seg->Z(),seg->TX(),seg->TY(),seg->DZ(),1,seg->W(),seg->ID());
+      run.GetView()->AddSegment(seg->X(),seg->Y(),seg->Z(),seg->TX(),seg->TY(),seg->DZ(),1,(int)(seg->W()),seg->ID());
     }
     run.AddView();
     run.GetView()->Clear();
@@ -482,7 +482,7 @@ void EdbDataStore::SavePlateToRaw(int PID,const char* dir,int id){
     run.GetView()->SetNframes(0,1);
     for(int ns=0;ns<pbot->N();++ns){
       seg=pbot->GetSegment(ns);
-      run.GetView()->AddSegment(seg->X(),seg->Y(),seg->Z(),seg->TX(),seg->TY(),seg->DZ(),2,seg->W(),seg->ID());
+      run.GetView()->AddSegment(seg->X(),seg->Y(),seg->Z(),seg->TX(),seg->TY(),seg->DZ(),2,(int)(seg->W()),seg->ID());
     }
     run.AddView();
     printf("close\n");
